Minimal support for layers
Added the main base for layers in core, now need better UI/UX. For now
we can create a layer above with `Y` and switch between first and second
layers with keys 1 and 2. All the activity happens in the active layer,
including saving. Probably there's a lot of bugs related to layers not
being synchronized, and between core layers and UI layer textures.