Save and load project done on frontend
Before the State was responsible for creating and loading project files,
now this is done in the front-end layer, since details of how the
project file should be structured should be left up to the frontend to
decide. The only thing done by State is to serialize and deserialize
itself to bytes.