Performance improvement generating image preview
The preview image was being generated with the `blended` function, which
calculates pixel by pixel on CPU. Now we draw the layer textures stacked on
top of each other via egui. There are still improvements to be done,
specially by not generating the preview when the preview window is closed, and
generally controlling in a smarter way when we need the preview to be
updated (we can do it less often).