Draw the preview image using macroquad
We use some features of egui to get the rectangle of the screen where
the image should be drawn, and also the offset of the scrollarea. With
this, we are able to draw all layers in a similar way as we do with the
canvas. A huge benefit of this is that now working with large images
does not have a big performance penalty.